Finished 0.1.0 implementation.
Split the PreferenceGroup class into two files using the partial keyword. The associated object logic is verbose and separating it from the core logic provides for better readability.
Updated the readme to provide diagrams of the structure of the package and to list what types are supported.
